+++ kwcoco/demo/toypatterns.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+from .demo.toypatterns import *  # NOQA
+import warnings
+warnings.warn('kwcoco.toypatterns is deprecated, use kwcoco.demo.toypatterns instead',
+              DeprecationWarning)